+++ src/main/java/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 public class Calculator {
 	public long calculate(String expression) {
