+++ src/main/java/calculator/Operator.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 public interface Operator {
 	long operation(long a, long b);
