+++ src/main/java/calculator/OperatorFactory.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package calculator;
 
 import java.util.HashMap;
 import java.util.Map;
