+++ src/main/java/racinggame/MoveStrategy.java
@@ -1,4 +1,4 @@
-package carracinggame;
+package racinggame;
 
 public interface MoveStrategy {
 	boolean isPossibleMove();
