+++ src/test/java/racinggame/CarTest.java
@@ -1,10 +1,12 @@
-package carracinggame;
+package racinggame;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import racinggame.domain.Car;
 
 class CarTest {
 	Car car;
